Changing some UInt16 variables to UInt8 to save resources.

+++ processors/renesas/src/eval/adi/hal_adi.c
@@ -78,10 +78,10 @@
 
 
 
-extern UInt16 rx_rf_idx;
-extern UInt16 rx_rf_count;
-extern UInt16 tx_rf_len;
-extern UInt16 tx_rf_idx;
+extern UInt8 rx_rf_idx;
+extern UInt8 rx_rf_count;
+extern UInt8 tx_rf_len;
+extern UInt8 tx_rf_idx;
 extern const UInt8 * tx_rf_data;
 extern UInt8 bit_mask;
